pyproject.toml
[project]
name = "rl-explainer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai (>=1.21.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
